--- llvm/lib/Frontend/Offloading/AMDGPU/CMakeLists.txt
@@ -1,8 +1,0 @@
-add_llvm_component_library(LLVMFrontendOffloadingAMDGPU
-  ObjectUtilities.cpp
-
-  LINK_COMPONENTS
-  Support
-  BinaryFormat
-  Object
-)
