+++ llvm/lib/Frontend/Offloading/CMakeLists.txt
@@ -1,5 +1,3 @@
-add_subdirectory(AMDGPU)
-
 add_llvm_component_library(LLVMFrontendOffloading
   Utility.cpp
   OffloadWrapper.cpp
@@ -13,6 +11,7 @@ add_llvm_component_library(LLVMFrontendOffloading
   LINK_COMPONENTS
   Core
   BinaryFormat
+  Object
   Support
   TransformUtils
   TargetParser
